refactor: Change ItemMain* to ItemMainRenderable* and use ItemMain* for non-renderable items (like troughs).
