+++ VisualPinball.Unity/VisualPinball.Unity/VPT/IItemMainRenderableAuthoring.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9001206d60c12264f9454a1da2928b9a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
