+++ VisualPinball.Unity/VisualPinball.Unity/VPT/ISwitchAuthoring.cs
@@ -14,7 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-
 using VisualPinball.Engine.Game;
 
 namespace VisualPinball.Unity
