+++ VisualPinball.Unity/VisualPinball.Unity/VPT/ItemMainRenderableAuthoring.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dbbfe725a97483b449275bc9d63c4f86
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
