+++ VisualPinball.Unity/VisualPinball.Unity/VPT/Trough/TroughExtensions.cs
@@ -24,7 +24,9 @@ namespace VisualPinball.Unity
 	{
 		public static ConvertedItem SetupGameObject(this Engine.VPT.Trough.Trough trough, GameObject obj)
 		{
+
 			var mainAuthoring = obj.AddComponent<TroughAuthoring>().SetItem(trough);
+			obj.GetComponentInParent<Player>().RegisterTrough(trough, obj);
 			return new ConvertedItem(mainAuthoring);
 		}
 	}
